subscription-tracker-backend/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
paddlepaddle==2.6.1
paddleocr==2.7.3
rapidfuzz==3.9.3
Pillow==10.3.0
